product cart functionality

+++ FrontEnd/src/app/product/components/product-list/product.component.ts
@@ -3,6 +3,7 @@ import { ProductService } from "src/app/shared-data/services/product.service";
 import { Router } from "@angular/router";
 import { Product } from "src/app/shared-data/models/product.model";
 import { CommonService } from "./../../../shared-data/services/common.service";
+import { CartService } from 'src/app/shared-data/services/cart.service';
 
 @Component({
   selector: "app-product",
@@ -20,7 +21,8 @@ export class ProductComponent implements OnInit {
   constructor(
     private router: Router,
     private productService: ProductService,
-    private commonService: CommonService
+    private commonService: CommonService,
+    private cartService:CartService
   ) {}
 
   ngOnInit() {
@@ -99,17 +101,28 @@ export class ProductComponent implements OnInit {
   }
 
   delete(product: Product) {
-    this.productService.deleteProduct(product._id).subscribe(response => {
-      if (response && response["result"]["success"] === true) {
-        this.commonService.showMessage(
-          "success",
-          "Product deleted successfully"
-        );
-        this.rows = this.rows.filter(_product => _product._id !== product._id);
-      } else {
-        this.commonService.showMessage("error", "Product can not be  deleted");
-      }
-    });
+    if (this.isProductAlreadyInCart(product._id) > -1) {
+      this.commonService.showMessage("error", "Product already in Cart . Product  can not be  deleted");
+    } else {
+      this.productService.deleteProduct(product._id).subscribe(response => {
+        if (response && response["result"]["success"] === true) {
+          this.commonService.showMessage(
+            "success",
+            "Product deleted successfully"
+          );
+          this.rows = this.rows.filter(_product => _product._id !== product._id);
+        } else {
+          this.commonService.showMessage("error", "Product can not be  deleted");
+        }
+      });
+
+    }
+
+  }
+
+  isProductAlreadyInCart(productId) {
+   const isExist =  this.cartService.itemExistIncart(productId);
+   return isExist;
   }
 
   navigateToCreate() {
